+++ lib/src/internal/plugin.dart
@@ -141,7 +141,7 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
     required PMPathFilter pathFilterOption,
   }) async {
     if (onlyAll) {
-      assert(hasAll, 'If only is true, then the hasAll must be not null.');
+      hasAll = true;
     }
     filterOption ??= FilterOptionGroup();
     final result = await _channel.invokeMethod<Map>(
@@ -350,7 +350,9 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
     required String? title,
     String? desc,
     String? relativePath,
+    int? orientation,
   }) async {
+    _throwIfOrientationInvalid(orientation);
     final Map<dynamic, dynamic>? result = await _channel.invokeMethod(
       PMConstants.mSaveImage,
       <String, dynamic>{
@@ -358,6 +360,7 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
         'title': title,
         'desc': desc ?? '',
         'relativePath': relativePath,
+        'orientation': orientation,
         ...onlyAddPermission,
       },
     );
@@ -375,11 +378,12 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
     required String title,
     String? desc,
     String? relativePath,
+    int? orientation,
   }) async {
+    _throwIfOrientationInvalid(orientation);
     final File file = File(path);
     if (!file.existsSync()) {
-      assert(false, 'File must exists.');
-      return null;
+      throw ArgumentError('$path does not exists');
     }
     final Map<dynamic, dynamic>? result = await _channel.invokeMethod(
       PMConstants.mSaveImageWithPath,
@@ -388,6 +392,7 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
         'title': title,
         'desc': desc ?? '',
         'relativePath': relativePath,
+        'orientation': orientation,
         ...onlyAddPermission,
       },
     );
@@ -405,10 +410,11 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
     required String? title,
     String? desc,
     String? relativePath,
+    int? orientation,
   }) async {
+    _throwIfOrientationInvalid(orientation);
     if (!file.existsSync()) {
-      assert(false, 'File must exists.');
-      return null;
+      throw ArgumentError('${file.path} does not exists');
     }
     final Map<dynamic, dynamic>? result = await _channel.invokeMethod(
       PMConstants.mSaveVideo,
@@ -417,6 +423,7 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
         'title': title,
         'desc': desc ?? '',
         'relativePath': relativePath,
+        'orientation': orientation,
         ...onlyAddPermission,
       },
     );
@@ -521,11 +528,9 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
     AssetPathEntity pathEntity,
   ) async {
     if (pathEntity.isAll) {
-      assert(
-        pathEntity.isAll,
-        "You can't copy the asset into the album containing all the pictures.",
+      throw ArgumentError(
+        'Cannot copy into the album containing all the assets.',
       );
-      return null;
     }
     final Map<dynamic, dynamic>? result = await _channel.invokeMethod(
       PMConstants.mCopyAsset,
@@ -573,7 +578,9 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
     if (PlatformUtils.isOhos) {
       return;
     }
-    assert(ids.isNotEmpty);
+    if (ids.isEmpty) {
+      throw ArgumentError('Empty IDs are not allowed');
+    }
     return _channel.invokeMethod(
       PMConstants.mRequestCacheAssetsThumb,
       <String, dynamic>{
@@ -584,7 +591,6 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
   }
 
   Future<void> presentLimited(RequestType type) async {
-    assert(Platform.isIOS || Platform.isAndroid);
     if (Platform.isIOS || Platform.isAndroid) {
       return _channel.invokeMethod(PMConstants.mPresentLimited, {
         'type': type.value,
@@ -593,12 +599,6 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
   }
 
   Future<String?> getMimeTypeAsync(AssetEntity entity) async {
-    assert(
-      Platform.isAndroid ||
-          Platform.isIOS ||
-          Platform.isMacOS ||
-          PlatformUtils.isOhos,
-    );
     if (Platform.isAndroid || PlatformUtils.isOhos) {
       return entity.mimeType;
     }
@@ -616,7 +616,6 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin, OhosPlugin {
     RequestType type = RequestType.common,
   }) {
     final filter = filterOption ?? PMFilter.defaultValue();
-
     return _channel.invokeMethod<int>(PMConstants.mGetAssetCount, {
       'type': type.value,
       'option': filter.toMap(),
@@ -676,14 +675,12 @@ mixin IosPlugin on BasePlugin {
   }) async {
     assert(Platform.isIOS || Platform.isMacOS);
     if (!imageFile.existsSync()) {
-      assert(false, 'File of the image must exist.');
-      return null;
+      throw ArgumentError('The image file does not exists');
     }
     if (!videoFile.existsSync()) {
-      assert(false, 'videoFile must exists.');
-      return null;
+      throw ArgumentError('The video file does not exists');
     }
-    final Map<dynamic, dynamic>? result = await _channel.invokeMethod(
+    final result = await _channel.invokeMethod(
       PMConstants.mSaveLivePhoto,
       <String, dynamic>{
         'imagePath': imageFile.absolute.path,
@@ -694,13 +691,7 @@ mixin IosPlugin on BasePlugin {
         ...onlyAddPermission,
       },
     );
-    if (result == null) {
-      return null;
-    }
-    return ConvertUtils.convertMapToAsset(
-      result.cast<String, dynamic>(),
-      title: title,
-    );
+    return ConvertUtils.convertMapToAsset(result.cast(), title: title);
   }
 
   Future<AssetPathEntity?> iosCreateAlbum(
@@ -830,4 +821,18 @@ mixin OhosPlugin on BasePlugin {
     }
     return result ?? <String>[];
   }
+}
+
+void _throwIfOrientationInvalid(int? value) {
+  if (value == null ||
+      value == 0 ||
+      value == 90 ||
+      value == 180 ||
+      value == 270) {
+    return;
+  }
+  throw ArgumentError(
+    'The given orientation is invalid, '
+    'allowed values are 0, 90, 180, 270, and null.',
+  );
 }